+++ NEWS.md
@@ -1,8 +1,9 @@
-# mirai 0.8.2.9002 (development)
+# mirai 0.8.2.9003 (development)
 
 * `dispatcher()` gains the argument 'token' for appending a unique token to each URL the dispatcher listens at.
 * `saisei()` implemented to regenerate the token used by a given dispatcher socket.
 * Local IPC socket URLs revised to use unique tokens.
+* Daemons status matrix headers updated to "online", "instance", "assigned", and "complete".
 * Requires nanonext >= [0.8.2].
 * Internal performance enhancements.